+++ core-ng/src/main/java/core/framework/impl/template/fragment/URLFragment.java
@@ -1,6 +1,5 @@
 package core.framework.impl.template.fragment;
 
-import core.framework.impl.template.CDNManager;
 import core.framework.impl.template.TemplateContext;
 import core.framework.impl.template.TemplateMetaContext;
 import core.framework.impl.template.expression.ExpressionBuilder;
@@ -11,13 +10,13 @@ import org.slf4j.LoggerFactory;
 /**
  * @author neo
  */
-public class CDNFragment implements Fragment {
-    private final Logger logger = LoggerFactory.getLogger(CDNFragment.class);
+public class URLFragment implements Fragment {  // this is for dynamic href only, static href won't be touched during compilation
+    private final Logger logger = LoggerFactory.getLogger(URLFragment.class);
 
     private final ExpressionHolder expression;
     private final String location;
 
-    public CDNFragment(String expression, TemplateMetaContext context, String location) {
+    public URLFragment(String expression, TemplateMetaContext context, String location) {
         this.expression = new ExpressionBuilder(expression, context, location).build();
         this.location = location;
     }
@@ -25,15 +24,15 @@ public class CDNFragment implements Fragment {
     @Override
     public void process(StringBuilder builder, TemplateContext context) {
         String url = String.valueOf(expression.eval(context));
-        builder.append(cdn(url, context.cdn));
+        builder.append(sanitize(url));
     }
 
-    private String cdn(String url, CDNManager cdn) {
-        //TODO: better way to prevent XSS?
+    private String sanitize(String url) {
+        //TODO: better way to prevent XSS? href should be more strict than cdn
         if (url.contains("javascript:") || url.contains("<") || url.contains(">") || url.contains(" ")) {
             logger.warn("illegal url detected, url={}, location={}", url, location);
             return "";
         }
-        return cdn.url(url);
+        return url;
     }
 }